Checking stream mode in Event#modified_files.
The mtime mode is the default and is the same as before, checking if any file's mtime is greater than or equal to the last event time.
The cache mode will check the stream's cache to see if any file's mtime or size has changed. (The checks happen explicitly because File::Stat#<=> only compares mtime).